+++ src/longmi/impute/__init__.py
@@ -1,9 +1,12 @@
 from .base import BaseFit, BaseImputer
+from .bernoulli import BernoulliFit, BernoulliImputer
 from .gaussian import JointGaussianFit, JointGaussianImputer
 from .negbin import NegativeBinomialFit, NegativeBinomialImputer
 
 __all__ = [
     "BaseFit",
+    "BernoulliFit",
+    "BernoulliImputer",
     "BaseImputer",
     "JointGaussianFit",
     "JointGaussianImputer",